Relocate ISelectedState + extract ITreeNode into headless Gum.Presentation (Phase 3, ADR-0005)
Continues ADR-0005 Phase 3 (after #3345/#3346). Moves the ISelectedState contract into the headless Gum.Presentation assembly, preserving its Gum.ToolStates namespace so consumers are unchanged. The concrete SelectedState implementation stays in the WPF shell.

Prerequisite, same PR: ITreeNode (framework-clean contract: object? Tag, FilePath GetFullFilePath(), ITreeNode? Parent, string Text/FullPath, Expand()) was declared inside the WinForms-coupled Gum/Managers/DragDropManager.cs. Extracted verbatim into Tools/Gum.Presentation/Managers/ITreeNode.cs, keeping the Gum.Managers namespace; its concrete tree-node implementation stays in the shell.

Also drops the unused 'using System.Windows.Forms;' from ISelectedState (the only WinForms reference in the file). ISelectedState's default-interface-method bodies construct only headless types (RecursiveVariableFinder, ElementWithState), so they compile in Gum.Presentation.

Behavior-preserving. Proof: Gum.Presentation builds standalone (0 errors; no WinForms/WPF reference in the csproj) and GumToolUnitTests is green (997 passed, 4 skipped).
